requirements.txt
# python 3.11.6
llama_cpp_python>=0.2.7
# pip install --index-url https://test.pypi.org/simple/ pymeteor
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
cloudpathlib==0.16.0
confection==0.1.3
cymem==2.0.8
distro==1.8.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.0/en_core_web_lg-3.7.0-py3-none-any.whl#sha256=708da1110fbe1163d059de34a2cbedb1db65c26e1e624ca925897a2711cb7d77
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.3
multidict==6.0.4
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.2
openai==0.28.0
packaging==23.2
pluggy==1.3.0
preshed==3.0.9
pydantic==2.5.1
pydantic_core==2.14.3
pymeteor==0.0.1
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
scipy==1.11.3
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlite-fts4==1.0.3
sqlite-utils==3.35.2
srsly==2.4.8
tabulate==0.9.0
thinc==8.2.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
wasabi==1.1.2
weasel==0.3.4
yarl==1.9.2
pandas==2.1.3
